unikraft/app: Split commands in types
